requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.6.0
django-countries==7.6.1
django-environ==0.12.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
mysqlclient==2.2.7
phonenumbers==8.13.54
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
